app/src/main/java/com/dicoding/moviecatalogue/ui/detail/movie/DetailMovieViewModel.kt
package com.dicoding.moviecatalogue.ui.detail.movie

import androidx.lifecycle.LiveData
import androidx.lifecycle.ViewModel
import com.dicoding.moviecatalogue.data.DataRepository
import com.dicoding.moviecatalogue.data.source.local.MovieEntity

class DetailMovieViewModel(private val dataRepository: DataRepository) : ViewModel() {
    private lateinit var movieId: String

    fun setSelectedMovie(movieId: String) {
        this.movieId = movieId
    }

    fun getMovie(): LiveData<MovieEntity> = dataRepository.getMovieById(movieId)
}